moving actions inside each directive
