+++ lib/js/interface/content.js
@@ -1,6 +1,9 @@
+/*global require*/
 import angular from 'angular';
 import npmApi from '../npm/npm-api.js';
-const moduleName = 'npm-ui.content';
+const moduleName = 'npm-ui.content'
+  , fileSystem = require('fs')
+  , Path = require('path');
 
 angular.module(moduleName, [
   npmApi
@@ -44,7 +47,7 @@ angular.module(moduleName, [
     unregisterLeftBarDeleteProjectListener();
   });
 })
-.directive('npmTabs', /*@ngInject*/ function npmTabs($rootScope, $log, npmGlobal, npm) {
+.directive('npmTabs', /*@ngInject*/ function npmTabs($rootScope, $log, npmGlobal, npm, errorsService, loadingFactory, notificationFactory) {
   return (scope, element, attrs) => {
 
     const tabIdentifierPath = attrs.npmTabId
@@ -71,11 +74,11 @@ angular.module(moduleName, [
           scope.showLoadingSelectedRow = false;
         }
       })
-      , unregisterInstallVersionPackageListener = $rootScope.$on('top-bar:installing-version-package', () => {
+      , unregisterInstallVersionPackageListener = $rootScope.$on('content:installing-version-package', () => {
           //while selected package and clicked "install release" and clicked prompt dialog button "ok" (while literally installing package new release)
           scope.showLoadingSelectedRow = true;
         })
-      , unregisterInstallVersionPackageErrorListener = $rootScope.$on('top-bar:installing-version-package-error', () => {
+      , unregisterInstallVersionPackageErrorListener = $rootScope.$on('content:installing-version-package-error', () => {
           //while selected package and clicked "install" + version, if no version available it shows an error message and we must remove active loading status from that package row in table
           scope.$apply(() => {
             scope.showLoadingSelectedRow = false;
@@ -85,6 +88,268 @@ angular.module(moduleName, [
         scope.packageInformations = [];
         scope.goBackHome = true;
       });
+
+    scope.installPackage = (pkgs, packageKind) => {
+      if (pkgs &&
+        pkgs.length > 0 &&
+        !scope.installingPackage) {
+
+        let promiseSequence = Promise.resolve()
+          , npmPromise
+          , choosedPackageKind
+          , installLaunch = () => {
+
+            pkgs.forEach(pkg => {
+
+              promiseSequence = promiseSequence.then(() => {
+                return npmPromise.catch(error => {
+                  errorsService.showErrorBox('Error', `Configuring npm for installing ${pkg.name}...`, error);
+                })
+                .then(npmInFolder => {
+                  $log.info(`Now installing ${pkg.name}...`);
+                  return npmInFolder.install({
+                    'name': pkg.name,
+                    'kind': choosedPackageKind
+                  }, pkg.version).then(() => {
+                    $log.info(`Installed ${pkg.name}...`);
+                  }).catch(error => {
+                    errorsService.showErrorBox('Error', `Installing ${pkg.name}: ${error}`);
+                  });
+                });
+              });
+            });
+
+            promiseSequence.then(() => {
+              loadingFactory.finished();
+              $log.info('Finished installing packages.');
+              notificationFactory.notify('Finished installing packages.');
+              scope.$apply(() => {
+                scope.installingPackage = false;
+              });
+              $rootScope.$emit('left-bar:select-project', {
+                'path': tabIdentifierPath
+              });
+            }).catch(error => {
+              scope.$apply(() => {
+                scope.installingPackage = false;
+              });
+              loadingFactory.finished();
+              errorsService.handleError('Error installing packages.', error);
+            });
+          };
+
+        scope.installingPackage = true;
+        loadingFactory.loading();
+
+        if (packageKind) {
+          choosedPackageKind = 'dev';
+        }
+
+        if (tabIdentifierPath === npmGlobal) {
+
+          npmPromise = npm.npmGlobal();
+        } else {
+
+          npmPromise = npm.npmInFolder(tabIdentifierPath);
+        }
+        //if not global install
+        if (tabIdentifierPath === npmGlobal) {
+          installLaunch();
+        } else {
+          //read package.json and create it if project is an empty folder
+          fileSystem.readFile(Path.join(tabIdentifierPath, 'package.json'), 'utf8', (err, packageJson) => {
+            if (err ||
+              !packageJson ||
+              (packageJson &&
+               JSON.parse(packageJson).length <= 0 &&
+               packageJson !== '{}')) {
+
+              $log.info(`Writing a package.json file in folder ${tabIdentifierPath}`);
+              //create package.json if not exist or if exist and not well formatted
+              fileSystem.writeFile(Path.join(tabIdentifierPath, 'package.json'), '{"name": "ndm-created-project", "description": "add-a-description", "license": "add-a-license", "respository": "add-a-repository"}', error => {
+                if (error) {
+                  installLaunch();
+                  $log.warn(`Unable to create package.json in: ${tabIdentifierPath}`);
+                } else {
+                  installLaunch();
+                  $log.info(`Initialized a package.json file in ${tabIdentifierPath}`);
+                }
+              });
+            } else {
+              installLaunch();
+            }
+          });
+        }
+      }
+    };
+
+    scope.updatePackage = currentSelectedPackages => {
+
+      let promiseSequence = Promise.resolve()
+        , npmPromise;
+
+      if (tabIdentifierPath === npmGlobal) {
+
+        npmPromise = npm.npmGlobal();
+      } else {
+
+        npmPromise = npm.npmInFolder(tabIdentifierPath);
+      }
+
+      currentSelectedPackages.forEach(pkg => {
+
+        promiseSequence = promiseSequence.then(() => {
+          return npmPromise.then(npmInFolder => {
+            $log.info(`Preparing to update ${pkg.name}...`);
+            return npmInFolder.update(pkg).then(() => {
+              $log.info(`Updated ${pkg.name}...`);
+            }).catch(error => {
+              errorsService.handleError(`Error updating ${pkg.name}...`, error);
+            });
+          }).catch(error => {
+            errorsService.handleError(`Error preparing to update ${pkg.name}...`, error);
+          });
+        });
+      });
+
+      promiseSequence.then(() => {
+        notificationFactory.notify('Finished updating selected packages.');
+        $log.info('Updated all the selected packages.');
+        $rootScope.$emit('left-bar:select-project', {
+          'path': tabIdentifierPath
+        });
+      }).catch(error => {
+        errorsService.handleError('Error updating package.', error);
+      });
+    };
+
+    scope.installLatest = currentSelectedPackages => {
+      let npmPromise
+        , promiseSequence = Promise.resolve();
+
+      if (tabIdentifierPath === npmGlobal) {
+
+        npmPromise = npm.npmGlobal();
+      } else {
+
+        npmPromise = npm.npmInFolder(tabIdentifierPath);
+      }
+
+      currentSelectedPackages.forEach(pkg => {
+
+        promiseSequence = promiseSequence.then(() => {
+          return npmPromise.then(npmInFolder => {
+            $log.info(`Preparing to update latest ${pkg.name}...`);
+            return npmInFolder.installLatest(pkg).then(() => {
+              $log.info(`Updated to latest ${pkg.name}...`);
+            }).catch(error => {
+              errorsService.handleError(`Error updating to latest ${pkg.name}...`, error);
+            });
+          }).catch(error => {
+            errorsService.handleError(`Error preparing to update to latest ${pkg.name}...`, error);
+          });
+        });
+      });
+
+      promiseSequence.then(() => {
+        $log.info('Updated all the selected packages to latest versions.');
+        notificationFactory.notify('Updated all the selected packages to latest versions.');
+        $rootScope.$emit('left-bar:select-project', {
+          'path': tabIdentifierPath
+        });
+      })
+      .catch(error => {
+        errorsService.handleError('Error updating to latest package version.', error);
+      });
+
+    };
+
+    scope.uninstallPackage = currentSelectedPackages => {
+
+      let promiseSequence = Promise.resolve()
+        , npmPromise;
+
+      if (tabIdentifierPath === npmGlobal) {
+
+        npmPromise = npm.npmGlobal();
+      } else {
+
+        npmPromise = npm.npmInFolder(tabIdentifierPath);
+      }
+
+      currentSelectedPackages.forEach(pkg => {
+
+        promiseSequence = promiseSequence.then(() => {
+          return npmPromise.then(npmInFolder => {
+            $log.info(`Preparing to uninstall ${pkg.name}...`);
+            return npmInFolder.rm(pkg).then(() => {
+              if (tabIdentifierPath === npmGlobal) {
+                $log.info(`Uninstalled ${pkg.name} from globals...`);
+              } else {
+                $log.info(`Uninstalled ${pkg.name} now pruning...`);
+                return npmInFolder.prune().then(() => {
+                  $log.info(`Uninstalled and Pruned ${pkg.name}...`);
+                });
+              }
+            }).catch(error => {
+              errorsService.handleError(`Error uninstalling ${pkg.name}...`, error);
+            });
+          }).catch(error => {
+            errorsService.handleError(`Error preparing to uninstall ${pkg.name}...`, error);
+          });
+        });
+      });
+
+      promiseSequence.then(() => {
+        $log.info('Uninstalled all the selected packages.');
+        notificationFactory.notify('Uninstalled all the selected packages.');
+        $rootScope.$emit('left-bar:select-project', {
+          'path': tabIdentifierPath
+        });
+      }).catch(error => {
+        errorsService.handleError('Error uninstalling package.', error);
+      });
+    };
+
+    scope.installVersionPackage = (currentSelectedPackage, specificVersion) => {
+      if (specificVersion &&
+        !scope.installingPackageVersion) {
+        scope.showSpecificVersionPrompt = false;
+        scope.installingPackageVersion = true;
+
+        let npmPromise;
+
+        if (tabIdentifierPath === npmGlobal) {
+
+          npmPromise = npm.npmGlobal();
+        } else {
+
+          npmPromise = npm.npmInFolder(tabIdentifierPath);
+        }
+
+        $rootScope.$emit('content:installing-version-package');
+        npmPromise.catch(error => {
+          scope.installingPackageVersion = false;
+          $rootScope.$emit('content:installing-version-package-error');
+          errorsService.showErrorBox('Error', `Error configuring npm for installing ${currentSelectedPackage.name}: ${error}`);
+        }).then(npmInFolder => {
+          npmInFolder.install(currentSelectedPackage, specificVersion)
+          .then(() => {
+            scope.installingPackageVersion = false;
+            notificationFactory.notify(`Finished installing ${currentSelectedPackage.name}@${specificVersion}`);
+            $rootScope.$emit('left-bar:select-project', {
+              'path': tabIdentifierPath
+            });
+          }).catch(error => {
+            scope.$apply(() => {
+              scope.installingPackageVersion = false;
+            });
+            $rootScope.$emit('content:installing-version-package-error');
+            errorsService.showErrorBox('Error', `Error installing ${currentSelectedPackage.name}@${specificVersion}: ${error}`);
+          });
+        });
+      }
+    };
 
     scope.loadPackagesInfos = packageName => {
       let folder = '/'; //not really needed probably
@@ -136,11 +401,8 @@ angular.module(moduleName, [
         scope.tableOrderBy.splice(scope.tableOrderBy.indexOf(by), 1);
       }
     };
-    //array of selected packages in table, we must init the array
-    scope.selectedPackages = [];
 
-    if (tabIdentifierPath) {
-
+    scope.listPackages = () => {
       if (typeof npm.npmGlobal === 'function' &&
         typeof npm.npmInFolder === 'function') {
 
@@ -187,6 +449,13 @@ angular.module(moduleName, [
       } else {
         $log.error('npm.npmGlobal or npm.infolder are not functions ready probably :/ :/ !? :/ content.js');
       }
+    };
+    //array of selected packages in table, we must init the array
+    scope.selectedPackages = [];
+
+    if (tabIdentifierPath) {
+
+      scope.listPackages();
     } else {
       $log.error('Path for tab identifier is missing');
     }
